pyproject.toml
[project]
name = "chatbot-stream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
    "streamlit>=1.54.0",
]
